+++ src/main/java/com/generation/blogpessoal/BlogPessoal/controller/TemaController.java
@@ -1,8 +1,12 @@
 package com.generation.blogpessoal.BlogPessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -12,14 +16,16 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.generation.blogpessoal.BlogPessoal.model.Tema;
 import com.generation.blogpessoal.BlogPessoal.repository.TemaRepository;
 
 @RestController
 @RequestMapping("/tema")
-@CrossOrigin("*")
+@CrossOrigin(origins = "*", allowedHeaders = "*")
 public class TemaController {
 
 	private @Autowired TemaRepository repository;
@@ -31,29 +37,37 @@ public class TemaController {
 
 	@GetMapping("/{id}")
 	public ResponseEntity<Tema> getById(@PathVariable long id) {
-		return repository.findById(id).map(resp -> ResponseEntity.ok(resp))
-				.orElse(ResponseEntity.notFound().build());
+		return repository.findById(id)
+				.map(resp -> ResponseEntity.ok(resp))
+				.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 	}
 
-	@GetMapping("/{descricao}")
+	@GetMapping("/descricao/{descricao}")
 	public ResponseEntity<List<Tema>> getByTemas(@PathVariable String descricao) {
 		return ResponseEntity.ok(repository.findAllByDescricaoContainingIgnoreCase(descricao));
 	}
 
 	@PostMapping("/new")
-	public ResponseEntity<Tema> newTemas(@RequestBody Tema newTemas) {
-		return ResponseEntity.status(201)
+	public ResponseEntity<Tema> newTemas(@Valid @RequestBody Tema newTemas) {
+		return ResponseEntity.status(HttpStatus.CREATED)
 				.body(repository.save(newTemas));
 	}
 
 	@PutMapping("/edit")
-	public ResponseEntity<Tema> editTemas(@RequestBody Tema editTemas) {
-		return ResponseEntity.status(200)
-				.body(repository.save(editTemas));
+	public ResponseEntity<Tema> editTemas(@Valid @RequestBody Tema editTemas) {
+		return repository.findById(editTemas.getId())
+				.map(resp -> ResponseEntity.status(HttpStatus.CREATED).body(repository.save(editTemas)))
+				.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 	}
 
+	@ResponseStatus(HttpStatus.NO_CONTENT)
 	@DeleteMapping("/delete/{id}")
 	public void deleteTemas(@PathVariable long id) {
+		Optional<Tema> tema = repository.findById(id);
+
+		if(tema.isEmpty())
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+		
 		repository.deleteById(id);
 	}
 }
